# Conflicts:
#	modules/platforms/dotnet/Apache.Ignite.Core.Tests/Binary/BinaryBuilderSelfTest.cs

+++ modules/core/src/main/java/org/apache/ignite/plugin/extensions/communication/MessageFactory.java
@@ -37,5 +37,5 @@ public interface MessageFactory extends Extension {
      * @param type Message type.
      * @return Message instance.
      */
-    @Nullable public Message create(byte type);
+    @Nullable public Message create(short type);
 }
